Update index.html
Minor edit to emphasise that it's fine for attendees to sign up for a flash talk if they just want to introduce themselves to other CogSci-adjacent folks but don't necessarily have a specific project they'd like to talk about.  I thought that requiring all attendees to talk about a project might be intimidating/offputting for undergraduate attendees, for example.